New Energy Bar

+++ Assets/Scripts/Base/PowerPlant.cs
@@ -5,7 +5,7 @@
 /// </summary>
 public class PowerPlant : MonoBehaviour {
     /// <summary>The energy production for an object</summary>
-    public float EnergyProduction;
+    public int EnergyProduction;
 
     /// <summary>Reference to BaseSwitcher to get EnergyPool on startup</summary>
     private BaseSwitcher baseSwitch;
